style: Fixed formatting for Customization screen as there was a duplicate.
